Align module folders to architecture
