+++ backend/modules/quiz/src/main/java/com/quizplatform/quiz/domain/service/DailyQuizServiceImpl.java
@@ -4,8 +4,8 @@ import com.quizplatform.quiz.domain.model.DifficultyLevel;
 import com.quizplatform.quiz.domain.model.Question;
 import com.quizplatform.quiz.domain.model.Quiz;
 import com.quizplatform.quiz.domain.model.QuizType;
-import com.quizplatform.quiz.domain.repository.QuestionRepository;
-import com.quizplatform.quiz.domain.repository.QuizRepository;
+import com.quizplatform.quiz.adapter.out.persistence.repository.QuestionRepository;
+import com.quizplatform.quiz.adapter.out.persistence.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
