+++ backend/modules/quiz/src/main/java/com/quizplatform/quiz/domain/service/QuizServiceImpl.java
@@ -3,8 +3,8 @@ package com.quizplatform.quiz.domain.service;
 import com.quizplatform.quiz.domain.model.Quiz;
 import com.quizplatform.quiz.domain.model.QuestionAttempt;
 import com.quizplatform.quiz.domain.model.QuizAttempt;
-import com.quizplatform.quiz.domain.repository.QuizAttemptRepository;
-import com.quizplatform.quiz.domain.repository.QuizRepository;
+import com.quizplatform.quiz.adapter.out.persistence.repository.QuizAttemptRepository;
+import com.quizplatform.quiz.adapter.out.persistence.repository.QuizRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
