--- backend/modules/user/src/main/java/com/quizplatform/user/infrastructure/http/dto/UserProfileUpdateRequest.java
@@ -1,26 +1,0 @@
-package com.quizplatform.user.infrastructure.http.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import jakarta.validation.constraints.Size;
-
-/**
- * 사용자 프로필 업데이트 요청 DTO
- *
- * @author 채기훈
- * @since JDK 21.0.6 Eclipse Temurin
- */
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class UserProfileUpdateRequest {
-    
-    @Size(min = 3, max = 50, message = "사용자명은 3자 이상 50자 이하여야 합니다")
-    private String username;
-    
-    private String profileImage;
-}
