+++ backend/modules/user/src/test/java/com/quizplatform/user/adapter/in/web/UserControllerTest.java
@@ -4,10 +4,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.quizplatform.user.domain.model.AuthProvider;
 import com.quizplatform.user.domain.model.User;
 import com.quizplatform.user.domain.model.UserRole;
-import com.quizplatform.user.domain.service.UserService;
-import com.quizplatform.user.infrastructure.http.UserController;
-import com.quizplatform.user.infrastructure.http.dto.UserCreationRequest;
-import com.quizplatform.user.infrastructure.http.dto.UserProfileUpdateRequest;
+import com.quizplatform.user.application.service.UserService;
+import com.quizplatform.user.adapter.in.web.UserController;
+import com.quizplatform.user.adapter.in.web.dto.UserCreationRequest;
+import com.quizplatform.user.adapter.in.web.dto.UserProfileUpdateRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
