+++ backend/modules/user/src/test/java/com/quizplatform/user/application/service/UserServiceTest.java
@@ -3,9 +3,9 @@ package com.quizplatform.user.application.service;
 import com.quizplatform.user.domain.model.AuthProvider;
 import com.quizplatform.user.domain.model.User;
 import com.quizplatform.user.domain.model.UserRole;
-import com.quizplatform.user.domain.repository.UserRepository;
-import com.quizplatform.user.domain.service.UserEventPublisher;
-import com.quizplatform.user.domain.service.UserServiceImpl;
+import com.quizplatform.user.adapter.out.persistence.repository.UserRepository;
+import com.quizplatform.user.adapter.out.event.UserEventPublisher;
+import com.quizplatform.user.application.service.UserServiceImpl;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
